2q KAK in quantum_info.synthesis, and quantum_info.random  (#2119)
* move kak to quantum_info. add Unitary definition

* remove MapperError in favor of TranspilerError

* add definition only for 2qubit Unitary

* add to __init__ of quantum_info sub packages

* add two_qubit_kak tests

* make quantum_info.random package for random state, unitary generation

* forgot to add random package

* lint and allow kak to accept ndarray or Operator

* allow compiler.transpile to not have backend. add pm.run().

* changelog
