+++ qiskit/mapper/exceptions.py
@@ -11,19 +11,6 @@ Exception for errors raised by mapper module.
 from qiskit.exceptions import QiskitError
 
 
-class MapperError(QiskitError):
-    """Base class for errors raised by mapper module."""
-
-    def __init__(self, *message):
-        """Set the error message."""
-        super().__init__(*message)
-        self.message = ' '.join(message)
-
-    def __str__(self):
-        """Return the message."""
-        return repr(self.message)
-
-
 class CouplingError(QiskitError):
     """Base class for errors raised by the coupling graph object."""
