+++ qiskit/quantum_info/operators/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2019, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Quantum Operators."""
+
+from .operator import Operator
+from .unitary import Unitary
+from .pauli import Pauli, pauli_group
+from .channel import Choi, SuperOp, Kraus, Stinespring, Chi, PTM
