+++ qiskit/quantum_info/random/__init__.py
@@ -5,6 +5,6 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Quantum States."""
+"""Methods for generating random quantum information objects."""
 
-from .states import basis_state, projector, purity
+from .utils import random_unitary, random_state, random_density_matrix
