+++ qiskit/quantum_info/synthesis/__init__.py
@@ -5,9 +5,6 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Utils for mapping."""
+"""State and Unitary synthesis methods."""
 
-from .compiling import two_qubit_kak, euler_angles_1q
-from .coupling import CouplingMap
-from .layout import Layout
-from .exceptions import CouplingError, LayoutError
+from .two_qubit_kak import two_qubit_kak, euler_angles_1q
