+++ test/python/test_qi.py
@@ -21,7 +21,7 @@ from qiskit.tools.qi.qi import shannon_entropy, entropy, mutual_information
 from qiskit.tools.qi.qi import choi_to_rauli, random_density_matrix
 from qiskit.tools.qi.qi import entanglement_of_formation, is_pos_def
 from qiskit.tools.qi.qi import __eof_qubit as eof_qubit
-from qiskit import QiskitError
+from qiskit.exceptions import QiskitError
 from qiskit.test import QiskitTestCase
